Modified the run command to allow for instant regeneration
The run command will now regenerate every time enter is pressed until the user enters a single "q" (case insensitive) or presses control-c. This makes for much faster iterating when testing design changes.

As a side-note, part of the motivation for this was that I found about every second use of the "run" command led to the server seemingly starting correctly but I couldn't connect on port 8000. Rerunning usually resolved the error. While this continues to happen, the failure to connect never occurs when using the new feature to regenerate the site - only when restarting it.