Foundation: Fix missing margin between sections in header
This change adds a bit of margin between each section within the header
of the built-in Foundation theme. It also increases the line height for
those links to make them render nicely on smaller screens when a larger
number of sections have been added, and makes the currently selected link
underlined.