Foundation+Section: Improve section rendering (#39)
This change adds a bit of margin between each section within the header
of the built-in Foundation theme. It also increases the line height for those
links to make them render nicely on smaller screens when a larger number
of sections have been added, and makes the currently selected link underlined.

It also makes the default title of all sections become a capitalized version
of their rawValue.