RSS: Add APIs for defining a prefix and suffix for an item’s body
This change makes it possible to define a prefix and/or suffix that will
get added to an item’s body when rendered within an RSS feed. This can,
for example, be used to provide specific intro paragraphs for RSS items,
to embed media players, and so on.

Just like the equivalent APIs for titles, these prefixes/suffixes are
rendered as-is and are not converted to HTML automatically.