RSS: Add option to customize an item’s link (#84)
This change makes it possible to specify a custom link for an item when
it appears in an RSS feed. This is useful for RSS items that link to
external resources, and enables those items to link directly to the target
URL. However, the website’s own internal URL is still used as a default
for each item’s GUID, in order to provide a stable identifier (as the
external URL might change). That can always be customized using `rss.guid`,
though.