Add initial support for Swift Concurrency (#117)
This change adds initial support for running certain publishing operations concurrently
using Swift's new built-in concurrency system. This doesn't involve any breaking changes
to the public API, apart from now requiring all Publish users to ensure that their HTML theme
and multi-item mutation code can be executed in a concurrent manner.

The following key changes were made as part of this:

- Bump the project's Swift version to `5.5`, and update the CLI's project template accordingly.
- Add CollectionConcurrencyKit as a new package dependency, which adds concurrent
   versions of the various sequence mapping functions that Publish makes heavy use of.
- Make `PublishingStep` use an `async` closure type as its body, which in turn enables
   asynchronous calls to be made within any step, both built-in and custom ones.
- Each publishing pipeline is still performed in sequence, just like before. So it's only the
   internal operations within each step that may now be parallelized.
- The top-level `Website.publish` APIs are still completely synchronous, and uses an internal
   `DispatchSemaphore` to await the completion of the website's publishing pipeline before
   returning.
- A new built-in `mutateAllItems(in sections:)` API has been added, to make it convenient
   to iterate over all items within a certain set of sections in a concurrent manner.
- `HTMLGenerator` now executes concurrently, meaning that it can output a large number
   of HTML files in parallel, rather than writing all files one-by-one.
- `MarkdownFileHandler` now also runs concurrently, and processes all of its Markdown files
   in parallel, rather than doing so sequentially.
- `RSSFeedGenerator` and `PodcastFeedGenerator` now generates their respective feeds
   concurrently.

Overall, this set of changes results in a 4x speed improvement when generating Swift by Sundell
using the M1 Max chip.

As part of this, Publish now requires macOS 12, to be able to use Swift Concurrency in a
predictable way. macOS version 12.0 has been out for quite a while now, so it's safe to require
it at this point.